Reorganize Integration Command tests
