+++ tests/php/Integration/Command/Format/Fixtures/bad-format.phel
