+++ tests/php/Integration/Command/Format/Fixtures/good-format.phel
