+++ tests/php/Integration/Command/Repl/Fixtures/exit-multiline-mode-with-ctrl-d.test
