+++ tests/php/Integration/Command/Repl/Fixtures/undefined-symbol.test
