+++ tests/php/Integration/Command/Run/Fixtures/test-script-without-ns.phel
