+++ tests/php/Integration/Command/Test/Fixtures/test-cmd-project-failure/composer.json
