+++ tests/php/Integration/Command/Test/Fixtures/test-cmd-project-success/composer.json
