+++ tests/php/Integration/Command/Test/TestCommandTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PhelTest\Integration\Command;
+namespace PhelTest\Integration\Command\Test;
 
 use Phel\Command\CommandFactory;
 use Phel\Command\CommandFactoryInterface;
